Extract MCP tool dispatch from app main
Refs #320
